refactor: reorganise folder structure
