+++ src/components/auth/SignUp.js
@@ -1,12 +1,12 @@
 import React, { useEffect, useState } from "react";
-import "../styles/Login.css";
+import "../../styles/Login.css";
 import { Link, useNavigate } from "react-router-dom";
 
-import { auth } from "../config/firebase";
+import { auth } from "../../config/firebase";
 import {
   createUserWithEmailAndPassword,
 } from "firebase/auth";
-import { getError } from "../logic/utils";
+import { getError } from "../../logic/utils";
 
 function SignUp() {
   const navigate = useNavigate();
